requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
backcall==0.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
distro==1.8.0
elevenlabs==0.2.26
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hume==0.4.1
idna==3.6
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.24.4
openai==1.3.5
opencv-python==4.8.1.78
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.41
pure-eval==0.2.2
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dotenv==1.0.0
requests==2.31.0
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
tqdm==4.66.1
traitlets==5.14.0
typing_extensions==4.8.0
urllib3==2.1.0
wcwidth==0.2.12
websockets==12.0
